*: make logging [log.Info] much more user-friendly
This is necessary to make it more obvious to users what is going on.

+++ cmd/umoci/config.go
@@ -279,21 +279,23 @@ func config(ctx *cli.Context) error {
 		return errors.Wrap(err, "commit mutated image")
 	}
 
-	log.WithFields(log.Fields{
-		"mediatype": newDescriptor.MediaType,
-		"digest":    newDescriptor.Digest,
-		"size":      newDescriptor.Size,
-	}).Infof("created new image")
+	log.Infof("new image manifest created: %s", newDescriptor.Digest)
 
-	// We have to clobber the old reference.
-	// XXX: Should we output some warning if we actually did remove an old
-	//      reference?
-	if err := engine.DeleteReference(context.Background(), tagName); err != nil {
-		return errors.Wrap(err, "delete old tag")
+	err = engine.PutReference(context.Background(), tagName, &newDescriptor)
+	if err == cas.ErrClobber {
+		// We have to clobber a tag.
+		log.Warnf("clobbering existing tag: %s", tagName)
+
+		// Delete the old tag.
+		if err := engine.DeleteReference(context.Background(), tagName); err != nil {
+			return errors.Wrap(err, "delete old tag")
+		}
+		err = engine.PutReference(context.Background(), tagName, &newDescriptor)
 	}
-	if err := engine.PutReference(context.Background(), tagName, &newDescriptor); err != nil {
+	if err != nil {
 		return errors.Wrap(err, "add new tag")
 	}
 
+	log.Infof("created new tag for image manifest: %s", tagName)
 	return nil
 }
