+++ cmd/umoci/new.go
@@ -64,7 +64,7 @@ func newImage(ctx *cli.Context) error {
 	// Create a new manifest.
 	log.WithFields(log.Fields{
 		"tag": tagName,
-	}).Infof("creating new manifest")
+	}).Debugf("creating new manifest")
 
 	// Create a new image config.
 	g := igen.New()
@@ -127,21 +127,24 @@ func newImage(ctx *cli.Context) error {
 		Size:      manifestSize,
 	}
 
-	log.WithFields(log.Fields{
-		"mediatype": descriptor.MediaType,
-		"digest":    descriptor.Digest,
-		"size":      descriptor.Size,
-	}).Infof("created new image")
+	log.Infof("new image manifest created: %s", descriptor.Digest)
 
-	// We have to clobber the old reference.
-	// XXX: Should we output some warning if we actually did remove an old
-	//      reference?
-	if err := engine.DeleteReference(context.Background(), tagName); err != nil {
-		return errors.Wrap(err, "delete old tag")
+	err = engine.PutReference(context.Background(), tagName, &descriptor)
+	if err == cas.ErrClobber {
+		// We have to clobber a tag.
+		log.Warnf("clobbering existing tag: %s", tagName)
+
+		// Delete the old tag.
+		if err := engine.DeleteReference(context.Background(), tagName); err != nil {
+			return errors.Wrap(err, "delete old tag")
+		}
+		err = engine.PutReference(context.Background(), tagName, &descriptor)
 	}
-	if err := engine.PutReference(context.Background(), tagName, &descriptor); err != nil {
+	if err != nil {
 		return errors.Wrap(err, "add new tag")
 	}
+
+	log.Infof("created new tag for image manifest: %s", tagName)
 
 	return nil
 }
